feat(rust): break up authenticator
