+++ implementations/rust/ockam/ockam_api/src/authenticator.rs
@@ -1,1 +1,2 @@
 pub mod direct;
+pub mod enrollment_tokens;
