+++ implementations/rust/ockam/ockam_command/src/project/ticket.rs
@@ -9,7 +9,8 @@ use std::time::Duration;
 use miette::{miette, IntoDiagnostic};
 use ockam::identity::Identifier;
 use ockam::Context;
-use ockam_api::authenticator::direct::{DirectAuthenticatorClient, TokenIssuerClient};
+use ockam_api::authenticator::direct::DirectAuthenticatorClient;
+use ockam_api::authenticator::enrollment_tokens::TokenIssuerClient;
 use ockam_api::cli_state::{CliState, StateDirTrait, StateItemTrait};
 use ockam_api::config::lookup::{ProjectAuthority, ProjectLookup};
 use ockam_api::DefaultAddress;
